remove audio control
