+++ Psyche/Assets/Prefabs/UI.prefab
@@ -2039,7 +2039,7 @@ GameObject:
   m_Component:
   - component: {fileID: 7823001781622809266}
   - component: {fileID: 6670206970594604876}
-  - component: {fileID: 3444753397153294863}
+  - component: {fileID: 8046082003159262787}
   m_Layer: 5
   m_Name: SliderVolSfx
   m_TagString: Untagged
@@ -2120,8 +2120,8 @@ MonoBehaviour:
   m_OnValueChanged:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 3444753397153294863}
-        m_TargetAssemblyTypeName: AudioControl, Assembly-CSharp
+      - m_Target: {fileID: 8046082003159262787}
+        m_TargetAssemblyTypeName: AudioManager, Assembly-CSharp
         m_MethodName: setSfxVolume
         m_Mode: 1
         m_Arguments:
@@ -2132,7 +2132,7 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 0
         m_CallState: 2
---- !u!114 &3444753397153294863
+--- !u!114 &8046082003159262787
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -2141,11 +2141,28 @@ MonoBehaviour:
   m_GameObject: {fileID: 1101621933854593256}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: bb20f0bab5ac10f41b09c4a8d26ef504, type: 3}
+  m_Script: {fileID: 11500000, guid: 9827f94284bfcf9478489ba72ef023aa, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  checkMark: {fileID: 2605753800641985410}
   slider: {fileID: 6670206970594604876}
+  uiVol: 0
+  musicVol: 0
+  sfxVol: 0
+  buttonClick: {fileID: 0}
+  uiAudioSources: []
+  backgroundMusic: {fileID: 0}
+  musicAudioSources: []
+  toolGRNS: {fileID: 0}
+  toolImager: {fileID: 0}
+  toolEMagnet: {fileID: 0}
+  toolThrusters: {fileID: 0}
+  playerJump: {fileID: 0}
+  playerHurt: {fileID: 0}
+  pickupElement: {fileID: 0}
+  pickupElementTungsten: {fileID: 0}
+  pickupTool: {fileID: 0}
+  checkpoint: {fileID: 0}
+  sfxAudioSources: []
 --- !u!1 &1101928457253581678
 GameObject:
   m_ObjectHideFlags: 0
@@ -8820,7 +8837,7 @@ RectTransform:
   m_Father: {fileID: 5188378539021892020}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 0.25, y: 1}
+  m_AnchorMax: {x: 0, y: 0}
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 10, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
@@ -10673,8 +10690,8 @@ RectTransform:
   m_Children: []
   m_Father: {fileID: 7962710141851774034}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
-  m_AnchorMin: {x: 0.25, y: 0}
-  m_AnchorMax: {x: 0.25, y: 1}
+  m_AnchorMin: {x: 0, y: 0}
+  m_AnchorMax: {x: 0, y: 0}
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 20, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
@@ -10726,7 +10743,7 @@ GameObject:
   m_Component:
   - component: {fileID: 9073134115627254300}
   - component: {fileID: 5669328167644140726}
-  - component: {fileID: 4405793089033215975}
+  - component: {fileID: 6574676012780896334}
   m_Layer: 5
   m_Name: SliderVolMusic
   m_TagString: Untagged
@@ -10807,8 +10824,8 @@ MonoBehaviour:
   m_OnValueChanged:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 4405793089033215975}
-        m_TargetAssemblyTypeName: AudioControl, Assembly-CSharp
+      - m_Target: {fileID: 6574676012780896334}
+        m_TargetAssemblyTypeName: AudioManager, Assembly-CSharp
         m_MethodName: setMusicVolume
         m_Mode: 1
         m_Arguments:
@@ -10819,7 +10836,7 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 0
         m_CallState: 2
---- !u!114 &4405793089033215975
+--- !u!114 &6574676012780896334
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -10828,11 +10845,28 @@ MonoBehaviour:
   m_GameObject: {fileID: 4336250559712014684}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: bb20f0bab5ac10f41b09c4a8d26ef504, type: 3}
+  m_Script: {fileID: 11500000, guid: 9827f94284bfcf9478489ba72ef023aa, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  checkMark: {fileID: 2605753800641985410}
   slider: {fileID: 5669328167644140726}
+  uiVol: 0
+  musicVol: 0
+  sfxVol: 0
+  buttonClick: {fileID: 0}
+  uiAudioSources: []
+  backgroundMusic: {fileID: 0}
+  musicAudioSources: []
+  toolGRNS: {fileID: 0}
+  toolImager: {fileID: 0}
+  toolEMagnet: {fileID: 0}
+  toolThrusters: {fileID: 0}
+  playerJump: {fileID: 0}
+  playerHurt: {fileID: 0}
+  pickupElement: {fileID: 0}
+  pickupElementTungsten: {fileID: 0}
+  pickupTool: {fileID: 0}
+  checkpoint: {fileID: 0}
+  sfxAudioSources: []
 --- !u!1 &4365923230534106531
 GameObject:
   m_ObjectHideFlags: 0
@@ -12838,7 +12872,7 @@ GameObject:
   m_Component:
   - component: {fileID: 1958126016903124576}
   - component: {fileID: 2642029444388704681}
-  - component: {fileID: 7562246366695072105}
+  - component: {fileID: 2677532683104193214}
   m_Layer: 5
   m_Name: SliderVolUi
   m_TagString: Untagged
@@ -12919,8 +12953,8 @@ MonoBehaviour:
   m_OnValueChanged:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 7562246366695072105}
-        m_TargetAssemblyTypeName: AudioControl, Assembly-CSharp
+      - m_Target: {fileID: 2677532683104193214}
+        m_TargetAssemblyTypeName: AudioManager, Assembly-CSharp
         m_MethodName: setUiVolume
         m_Mode: 1
         m_Arguments:
@@ -12931,7 +12965,7 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 0
         m_CallState: 2
---- !u!114 &7562246366695072105
+--- !u!114 &2677532683104193214
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -12940,11 +12974,28 @@ MonoBehaviour:
   m_GameObject: {fileID: 5262003178582905631}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: bb20f0bab5ac10f41b09c4a8d26ef504, type: 3}
+  m_Script: {fileID: 11500000, guid: 9827f94284bfcf9478489ba72ef023aa, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  checkMark: {fileID: 2605753800641985410}
   slider: {fileID: 2642029444388704681}
+  uiVol: 0
+  musicVol: 0
+  sfxVol: 0
+  buttonClick: {fileID: 0}
+  uiAudioSources: []
+  backgroundMusic: {fileID: 0}
+  musicAudioSources: []
+  toolGRNS: {fileID: 0}
+  toolImager: {fileID: 0}
+  toolEMagnet: {fileID: 0}
+  toolThrusters: {fileID: 0}
+  playerJump: {fileID: 0}
+  playerHurt: {fileID: 0}
+  pickupElement: {fileID: 0}
+  pickupElementTungsten: {fileID: 0}
+  pickupTool: {fileID: 0}
+  checkpoint: {fileID: 0}
+  sfxAudioSources: []
 --- !u!1 &5368865855472790935
 GameObject:
   m_ObjectHideFlags: 0
@@ -13752,7 +13803,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!224 &3739745722489845252
 RectTransform:
   m_ObjectHideFlags: 0
@@ -15676,8 +15727,8 @@ RectTransform:
   m_Children: []
   m_Father: {fileID: 6942531045086446441}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
-  m_AnchorMin: {x: 0.25, y: 0}
-  m_AnchorMax: {x: 0.25, y: 1}
+  m_AnchorMin: {x: 0, y: 0}
+  m_AnchorMax: {x: 0, y: 0}
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 20, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
@@ -18043,8 +18094,8 @@ RectTransform:
   m_Children: []
   m_Father: {fileID: 758790101232706279}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
-  m_AnchorMin: {x: 0.25, y: 0}
-  m_AnchorMax: {x: 0.25, y: 1}
+  m_AnchorMin: {x: 0, y: 0}
+  m_AnchorMax: {x: 0, y: 0}
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 20, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
@@ -20274,7 +20325,7 @@ RectTransform:
   m_Father: {fileID: 8087027586457339891}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 0.25, y: 1}
+  m_AnchorMax: {x: 0, y: 0}
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 10, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
@@ -22674,7 +22725,7 @@ RectTransform:
   m_Father: {fileID: 5661937759175808950}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 0.25, y: 1}
+  m_AnchorMax: {x: 0, y: 0}
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 10, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
