--- Psyche/Assets/Scripts/UI/AudioControl.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bb20f0bab5ac10f41b09c4a8d26ef504
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
